chore: eliminate churn from provider-management move-back
Remove pure reordering churn left over from the provider-management
module extraction + move-back. No functional change, ~400 lines of
diff removed vs feat/module-system.

- org_provider_keys / org_models: restored into organizations.ts
  (their original location) instead of a new provider-keys.ts file
- schema/types.ts: reverted reordering, kept OrgProviderKey/OrgModel
  type exports
- openapi/schemas.ts: restored OrgProviderKey/OrgModel/TestResult to
  their original position (before OrgProxy) instead of appended at end
- openapi/zod-schema-registry.ts: restored Models and Provider Keys
  imports + entries to their original positions
