+++ apps/api/src/openapi/schemas.ts
@@ -737,6 +737,78 @@ export const schemas = {
       createdAt: { type: ["string", "null"], format: "date-time" },
     },
   },
+  OrgProviderKey: {
+    type: "object",
+    required: ["id", "label", "api", "baseUrl", "source", "createdAt", "updatedAt"],
+    properties: {
+      id: { type: "string" },
+      label: { type: "string" },
+      api: { type: "string" },
+      baseUrl: { type: "string" },
+      source: { type: "string", enum: ["built-in", "custom"] },
+      createdBy: { type: ["string", "null"] },
+      createdAt: { type: "string", format: "date-time" },
+      updatedAt: { type: "string", format: "date-time" },
+    },
+  },
+  OrgModel: {
+    type: "object",
+    required: [
+      "id",
+      "label",
+      "api",
+      "baseUrl",
+      "modelId",
+      "enabled",
+      "isDefault",
+      "source",
+      "providerKeyId",
+      "createdAt",
+      "updatedAt",
+    ],
+    properties: {
+      id: { type: "string" },
+      label: { type: "string" },
+      api: { type: "string" },
+      baseUrl: { type: "string" },
+      modelId: { type: "string" },
+      input: { type: ["array", "null"], items: { type: "string" } },
+      contextWindow: { type: ["integer", "null"] },
+      maxTokens: { type: ["integer", "null"] },
+      reasoning: { type: ["boolean", "null"] },
+      enabled: { type: "boolean" },
+      isDefault: { type: "boolean" },
+      source: { type: "string", enum: ["built-in", "custom"] },
+      providerKeyId: {
+        type: "string",
+        description: "Provider key ID for API key credentials",
+      },
+      providerKeyLabel: { type: ["string", "null"], description: "Provider key label for display" },
+      cost: {
+        type: ["object", "null"],
+        description: "Cost per million tokens",
+        properties: {
+          input: { type: "number" },
+          output: { type: "number" },
+          cacheRead: { type: "number" },
+          cacheWrite: { type: "number" },
+        },
+      },
+      createdBy: { type: ["string", "null"] },
+      createdAt: { type: "string", format: "date-time" },
+      updatedAt: { type: "string", format: "date-time" },
+    },
+  },
+  TestResult: {
+    type: "object",
+    required: ["ok", "latency"],
+    properties: {
+      ok: { type: "boolean" },
+      latency: { type: "number", description: "Response time in milliseconds" },
+      error: { type: "string", description: "Error code if test failed" },
+      message: { type: "string", description: "Human-readable error message" },
+    },
+  },
   OrgProxy: {
     type: "object",
     required: ["id", "label", "enabled", "isDefault", "source", "createdAt", "updatedAt"],
@@ -844,78 +916,6 @@ export const schemas = {
           description: "Hint text shown before the user provides a value",
         },
       },
-    },
-  },
-  OrgProviderKey: {
-    type: "object",
-    required: ["id", "label", "api", "baseUrl", "source", "createdAt", "updatedAt"],
-    properties: {
-      id: { type: "string" },
-      label: { type: "string" },
-      api: { type: "string" },
-      baseUrl: { type: "string" },
-      source: { type: "string", enum: ["built-in", "custom"] },
-      createdBy: { type: ["string", "null"] },
-      createdAt: { type: "string", format: "date-time" },
-      updatedAt: { type: "string", format: "date-time" },
-    },
-  },
-  OrgModel: {
-    type: "object",
-    required: [
-      "id",
-      "label",
-      "api",
-      "baseUrl",
-      "modelId",
-      "enabled",
-      "isDefault",
-      "source",
-      "providerKeyId",
-      "createdAt",
-      "updatedAt",
-    ],
-    properties: {
-      id: { type: "string" },
-      label: { type: "string" },
-      api: { type: "string" },
-      baseUrl: { type: "string" },
-      modelId: { type: "string" },
-      input: { type: ["array", "null"], items: { type: "string" } },
-      contextWindow: { type: ["integer", "null"] },
-      maxTokens: { type: ["integer", "null"] },
-      reasoning: { type: ["boolean", "null"] },
-      enabled: { type: "boolean" },
-      isDefault: { type: "boolean" },
-      source: { type: "string", enum: ["built-in", "custom"] },
-      providerKeyId: {
-        type: "string",
-        description: "Provider key ID for API key credentials",
-      },
-      providerKeyLabel: { type: ["string", "null"], description: "Provider key label for display" },
-      cost: {
-        type: ["object", "null"],
-        description: "Cost per million tokens",
-        properties: {
-          input: { type: "number" },
-          output: { type: "number" },
-          cacheRead: { type: "number" },
-          cacheWrite: { type: "number" },
-        },
-      },
-      createdBy: { type: ["string", "null"] },
-      createdAt: { type: "string", format: "date-time" },
-      updatedAt: { type: "string", format: "date-time" },
-    },
-  },
-  TestResult: {
-    type: "object",
-    required: ["ok", "latency"],
-    properties: {
-      ok: { type: "boolean" },
-      latency: { type: "number", description: "Response time in milliseconds" },
-      error: { type: "string", description: "Error code if test failed" },
-      message: { type: "string", description: "Human-readable error message" },
     },
   },
 } as const;
