+++ packages/db/src/schema/organizations.ts
@@ -6,6 +6,7 @@ import {
   timestamp,
   boolean,
   uuid,
+  integer,
   jsonb,
   index,
   uniqueIndex,
@@ -122,6 +123,60 @@ export const orgProxies = pgTable(
   (table) => [
     index("idx_org_proxies_org_id").on(table.orgId),
     uniqueIndex("idx_org_proxies_one_default")
+      .on(table.orgId)
+      .where(sql`${table.isDefault} = true`),
+  ],
+);
+
+export const orgProviderKeys = pgTable(
+  "org_provider_keys",
+  {
+    id: uuid("id").defaultRandom().primaryKey(),
+    orgId: uuid("org_id")
+      .notNull()
+      .references(() => organizations.id, { onDelete: "cascade" }),
+    label: text("label").notNull(),
+    api: text("api").notNull(),
+    baseUrl: text("base_url").notNull(),
+    apiKeyEncrypted: text("api_key_encrypted").notNull(),
+    createdBy: text("created_by").references(() => user.id),
+    createdAt: timestamp("created_at").defaultNow().notNull(),
+    updatedAt: timestamp("updated_at").defaultNow().notNull(),
+  },
+  (t) => [index("idx_org_provider_keys_org_id").on(t.orgId)],
+);
+
+export const orgModels = pgTable(
+  "org_models",
+  {
+    id: uuid("id").defaultRandom().primaryKey(),
+    orgId: uuid("org_id")
+      .notNull()
+      .references(() => organizations.id, { onDelete: "cascade" }),
+    label: text("label").notNull(),
+    api: text("api").notNull(),
+    baseUrl: text("base_url").notNull(),
+    modelId: text("model_id").notNull(),
+    providerKeyId: uuid("provider_key_id")
+      .notNull()
+      .references(() => orgProviderKeys.id, {
+        onDelete: "cascade",
+      }),
+    input: jsonb("input"), // ["text", "image"] | null
+    contextWindow: integer("context_window"), // 200000 | null
+    maxTokens: integer("max_tokens"), // 16384 | null
+    reasoning: boolean("reasoning"), // true | null
+    cost: jsonb("cost"), // { input, output, cacheRead, cacheWrite } in $/M tokens | null
+    enabled: boolean("enabled").notNull().default(true),
+    isDefault: boolean("is_default").notNull().default(false),
+    source: text("source").notNull().default("custom"), // "built-in" | "custom"
+    createdBy: text("created_by").references(() => user.id),
+    createdAt: timestamp("created_at").defaultNow().notNull(),
+    updatedAt: timestamp("updated_at").defaultNow().notNull(),
+  },
+  (table) => [
+    index("idx_org_models_org_id").on(table.orgId),
+    uniqueIndex("idx_org_models_one_default")
       .on(table.orgId)
       .where(sql`${table.isDefault} = true`),
   ],
